pyproject.toml
[tool.poetry]
name = "minew-indoor-position"
version = "0.1.0"
description = ""
authors = ["paicaloid <paicaloid@gmail.com>"]
readme = "README.md"
packages = [{include = "minew_indoor_position"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^2.1.0"
numpy = "^2.1.1"
matplotlib = "^3.9.2"
redis = "^5.0.8"
ruff = "^0.6.3"
isort = "^5.13.2"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true